Misc: Fix Repo links, various refinements
Add logging to chooseAttack (not proper logging, just quick debug stuff)

Fix duplicate lock line in createAIAirplane, causing statics to be locked incorrectly

Move the TFAR SCR536 into its own addon for proper addon structure

Split ForkChangelog into ChangelogPlus and ChangelogUltimate

Updated README.md
