+++ A3A/addons/ultimate/CfgVehicles.hpp
@@ -1,0 +1,4 @@
+class Land_HelipadCircle_F;
+class Land_HelipadSquare_F;
+class A3AU_RebHelipad_Circle_F: Land_HelipadCircle_F {};
+class A3AU_RebHelipad_Square_F: Land_HelipadSquare_F {};
