+++ A3A/addons/ultimate/patches/cfgVehicles.hpp
@@ -25,7 +25,3 @@ class Prop_TFAR_SCR536 : Thing
     // editorCategory = "";
     // vehicleClass="";
 };
-class Land_HelipadCircle_F;
-class Land_HelipadSquare_F;
-class A3AU_RebHelipad_Circle_F: Land_HelipadCircle_F {};
-class A3AU_RebHelipad_Square_F: Land_HelipadSquare_F {};
