+++ A3A/addons/ultimate/script_component.hpp
@@ -1,2 +1,3 @@
 #define COMPONENT ultimate
 #include "\x\A3A\addons\core\Includes\script_mod.hpp"
+#define PATCHNAME(x) ADDON_ULTIMATE##_##x
